+++ packages/gatsby/cache-dir/__tests__/static-entry.js
@@ -2,7 +2,8 @@ import React from "react"
 import fs from "fs"
 const { join } = require(`path`)
 
-import DevelopStaticEntry from "../develop-static-entry"
+import ssrDevelopStaticEntry from "../ssr-develop-static-entry"
+import developStaticEntry from "../develop-static-entry"
 
 jest.mock(`fs`, () => {
   const fs = jest.requireActual(`fs`)
@@ -16,6 +17,19 @@ jest.mock(`gatsby/package.json`, () => {
     version: `2.0.0`,
   }
 })
+jest.mock(
+  `$virtual/ssr-sync-requires`,
+  () => {
+    return {
+      components: {
+        "page-component---src-pages-test-js": () => null,
+      },
+    }
+  },
+  {
+    virtual: true,
+  }
+)
 
 jest.mock(
   `$virtual/sync-requires`,
@@ -48,10 +62,10 @@ const MOCK_FILE_INFO = {
   }),
 }
 
-let StaticEntry
+let staticEntry
 beforeEach(() => {
   fs.readFileSync.mockImplementation(file => MOCK_FILE_INFO[file])
-  StaticEntry = require(`../static-entry`).default
+  staticEntry = require(`../static-entry`).default
 })
 
 const reverseHeadersPlugin = {
@@ -138,10 +152,69 @@ const fakeComponentsPluginFactory = type => {
 }
 
 describe(`develop-static-entry`, () => {
+  beforeEach(() => {
+    global.__PATH_PREFIX__ = ``
+    global.__BASE_PATH__ = ``
+    global.__ASSET_PREFIX__ = ``
+  })
+
+  test(`SSR: onPreRenderHTML can be used to replace headComponents`, done => {
+    global.plugins = [fakeStylesPlugin, reverseHeadersPlugin]
+
+    ssrDevelopStaticEntry(`/about/`, false, (_, html) => {
+      expect(html).toMatchSnapshot()
+      done()
+    })
+  })
+
+  test(`SSR: onPreRenderHTML can be used to replace postBodyComponents`, done => {
+    global.plugins = [
+      fakeComponentsPluginFactory(`Post`),
+      reverseBodyComponentsPluginFactory(`Post`),
+    ]
+
+    ssrDevelopStaticEntry(`/about/`, false, (_, html) => {
+      expect(html).toMatchSnapshot()
+      done()
+    })
+  })
+
+  test(`SSR: onPreRenderHTML can be used to replace preBodyComponents`, done => {
+    global.plugins = [
+      fakeComponentsPluginFactory(`Pre`),
+      reverseBodyComponentsPluginFactory(`Pre`),
+    ]
+
+    ssrDevelopStaticEntry(`/about/`, false, (_, html) => {
+      expect(html).toMatchSnapshot()
+      done()
+    })
+  })
+
+  test(`SSR: onPreRenderHTML adds metatag note for development environment`, done => {
+    ssrDevelopStaticEntry(`/about/`, false, (_, html) => {
+      expect(html).toContain(
+        `<meta name="note" content="environment=development"/>`
+      )
+      done()
+    })
+  })
+
+  test(`SSR: onPreRenderHTML adds metatag note for development environment after replaceHeadComponents`, done => {
+    global.plugins = [reverseHeadersPlugin]
+
+    ssrDevelopStaticEntry(`/about/`, false, (_, html) => {
+      expect(html).toContain(
+        `<meta name="note" content="environment=development"/>`
+      )
+      done()
+    })
+  })
+
   test(`onPreRenderHTML can be used to replace headComponents`, done => {
     global.plugins = [fakeStylesPlugin, reverseHeadersPlugin]
 
-    DevelopStaticEntry(`/about/`, (_, html) => {
+    developStaticEntry(`/about/`, (_, html) => {
       expect(html).toMatchSnapshot()
       done()
     })
@@ -153,7 +226,7 @@ describe(`develop-static-entry`, () => {
       reverseBodyComponentsPluginFactory(`Post`),
     ]
 
-    DevelopStaticEntry(`/about/`, (_, html) => {
+    developStaticEntry(`/about/`, (_, html) => {
       expect(html).toMatchSnapshot()
       done()
     })
@@ -165,14 +238,14 @@ describe(`develop-static-entry`, () => {
       reverseBodyComponentsPluginFactory(`Pre`),
     ]
 
-    DevelopStaticEntry(`/about/`, (_, html) => {
+    developStaticEntry(`/about/`, (_, html) => {
       expect(html).toMatchSnapshot()
       done()
     })
   })
 
   test(`onPreRenderHTML adds metatag note for development environment`, done => {
-    DevelopStaticEntry(`/about/`, (_, html) => {
+    developStaticEntry(`/about/`, (_, html) => {
       expect(html).toContain(
         `<meta name="note" content="environment=development"/>`
       )
@@ -183,7 +256,7 @@ describe(`develop-static-entry`, () => {
   test(`onPreRenderHTML adds metatag note for development environment after replaceHeadComponents`, done => {
     global.plugins = [reverseHeadersPlugin]
 
-    DevelopStaticEntry(`/about/`, (_, html) => {
+    developStaticEntry(`/about/`, (_, html) => {
       expect(html).toContain(
         `<meta name="note" content="environment=development"/>`
       )
@@ -210,7 +283,7 @@ describe(`static-entry sanity checks`, () => {
       const plugin = injectValuePlugin(`onPreRenderHTML`, methodName, null)
       global.plugins = [plugin, checkNonEmptyHeadersPlugin]
 
-      StaticEntry(`/about/`, (_, html) => {
+      staticEntry(`/about/`, (_, html) => {
         done()
       })
     })
@@ -222,7 +295,7 @@ describe(`static-entry sanity checks`, () => {
       ])
       global.plugins = [plugin, checkNonEmptyHeadersPlugin]
 
-      StaticEntry(`/about/`, (_, html) => {
+      staticEntry(`/about/`, (_, html) => {
         done()
       })
     })
@@ -231,7 +304,7 @@ describe(`static-entry sanity checks`, () => {
       const plugin = injectValuePlugin(`onPreRenderHTML`, methodName, [])
       global.plugins = [plugin, checkNonEmptyHeadersPlugin]
 
-      StaticEntry(`/about/`, (_, html) => {
+      staticEntry(`/about/`, (_, html) => {
         done()
       })
     })
@@ -240,7 +313,7 @@ describe(`static-entry sanity checks`, () => {
       const plugin = injectValuePlugin(`onPreRenderHTML`, methodName, [[], []])
       global.plugins = [plugin, checkNonEmptyHeadersPlugin]
 
-      StaticEntry(`/about/`, (_, html) => {
+      staticEntry(`/about/`, (_, html) => {
         done()
       })
     })
@@ -254,7 +327,7 @@ describe(`static-entry sanity checks`, () => {
       ])
       global.plugins = [plugin, checkNonEmptyHeadersPlugin]
 
-      StaticEntry(`/about/`, (_, html) => {
+      staticEntry(`/about/`, (_, html) => {
         done()
       })
     })
@@ -271,7 +344,7 @@ describe(`static-entry`, () => {
   test(`onPreRenderHTML can be used to replace headComponents`, done => {
     global.plugins = [fakeStylesPlugin, reverseHeadersPlugin]
 
-    StaticEntry(`/about/`, (_, html) => {
+    staticEntry(`/about/`, (_, html) => {
       expect(html).toMatchSnapshot()
       done()
     })
@@ -283,7 +356,7 @@ describe(`static-entry`, () => {
       reverseBodyComponentsPluginFactory(`Post`),
     ]
 
-    StaticEntry(`/about/`, (_, html) => {
+    staticEntry(`/about/`, (_, html) => {
       expect(html).toMatchSnapshot()
       done()
     })
@@ -295,14 +368,14 @@ describe(`static-entry`, () => {
       reverseBodyComponentsPluginFactory(`Pre`),
     ]
 
-    StaticEntry(`/about/`, (_, html) => {
+    staticEntry(`/about/`, (_, html) => {
       expect(html).toMatchSnapshot()
       done()
     })
   })
 
   test(`onPreRenderHTML does not add metatag note for development environment`, done => {
-    StaticEntry(`/about/`, (_, html) => {
+    staticEntry(`/about/`, (_, html) => {
       expect(html).not.toContain(
         `<meta name="note" content="environment=development"/>`
       )